# specs/001-heliolus-platform/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Heliolus Compliance Assessment Platform API
  description: API for compliance assessment, gap analysis, and vendor marketplace
  version: 1.0.0
  contact:
    name: Heliolus Support
    email: support@heliolus.com
servers:
  - url: https://api.heliolus.com/v1
    description: Production
  - url: https://staging-api.heliolus.com/v1
    description: Staging
  - url: http://localhost:3001/v1
    description: Development
security:
  - bearerAuth: []
...
paths:
...
  /organizations/{id}/documents:
    get:
      tags: [Documents]
      summary: List organization documents
      operationId: listDocuments
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        200:
          description: Document list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Document'

    post:
      tags: [Documents]
      summary: Upload document
      operationId: uploadDocument
      parameters:
        - $ref: '#/components/parameters/id'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                documentType:
                  $ref: '#/components/schemas/DocumentType'
      responses:
        201:
          description: Document uploaded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        400:
          $ref: '#/components/responses/BadRequest'
  /documents/{id}:
    get:
      tags: [Documents]
      summary: Get document
      operationId: getDocument
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        200:
          description: Document details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        404:
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Documents]
      summary: Delete document
      operationId: deleteDocument
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        204:
          description: Document deleted
        404:
          $ref: '#/components/responses/NotFound'

  # Templates
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    id:
      name: id
      in: path
      required: true
      schema:
        type: string
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      required: [message]
      properties:
        message:
          type: string
        code:
          type: string
        details:
          type: object
...
    Document:
      type: object
      properties:
        id:
          type: string
        filename:
          type: string
        originalName:
          type: string
        mimeType:
          type: string
        size:
          type: integer
        documentType:
          $ref: '#/components/schemas/DocumentType'
        createdAt:
          type: string
          format: date-time
    DocumentType:
      type: string
      enum: [POLICY, ANNUAL_REPORT, COMPLIANCE_CERT, AUDIT_REPORT, OTHER]
...